fix: move identity mapping to "Input" section
